correction NullPointeur Exception

+++ src/gpi/view/SiteOverviewController.java
@@ -71,7 +71,7 @@ public class SiteOverviewController {
                     public void handle(MouseEvent arg0) {
                         System.out.println(((Label)(tempo.getBottom())).getId());
                         String id = ((Label)(tempo.getBottom())).getId();
-                        mainApp.setCritere((Integer)(Integer.parseInt(id)));
+                        mainApp.setCritere((Integer.parseInt(id)));
                         mainApp.changerTab("Type");
                     }
                 });
